Bootstrap implementation of QueryableCollection for normal collection
